authenticated get, put, delete

+++ routers/todos.py
@@ -31,13 +31,17 @@ class TodoRequest(BaseModel):
 
 
 @router.get("/", status_code=status.HTTP_200_OK)
-async def read_all(db: db_dependency):
-    return db.query(Todos).all()
+async def read_all(user: user_dependencies, db: db_dependency):
+    if not user:
+        raise HTTPException(status_code=404, detail="Authentication Failed")
+    return db.query(Todos).filter(Todos.owner_id == int(user.get('id'))).all()
 
 
 @router.get("/todo/{todo_id}", status_code=status.HTTP_200_OK)
-async def read_todo(db: db_dependency, todo_id: int = Path(ge=1)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def read_todo(user: user_dependencies, db: db_dependency, todo_id: int = Path(ge=1)):
+    if not user:
+        raise HTTPException(status_code=404, detail="Authentication Failed")
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == int(user.get('id'))).first()
     if todo_model:
         return todo_model
     else:
@@ -54,8 +58,10 @@ async def create_todo(user: user_dependencies, db: db_dependency, todo_request: 
 
 
 @router.put("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def update_todo(db: db_dependency, todo_request: TodoRequest, todo_id: int = Path(ge=1)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def update_todo(user: user_dependencies, db: db_dependency, todo_request: TodoRequest, todo_id: int = Path(ge=1)):
+    if not user:
+        raise HTTPException(status_code=404, detail="Authentication Failed")
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == int(user.get('id'))).first()
     if todo_model:
         todo_model.title = todo_request.title
         todo_model.description = todo_request.description
@@ -68,8 +74,10 @@ async def update_todo(db: db_dependency, todo_request: TodoRequest, todo_id: int
 
 
 @router.delete("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_todo(db: db_dependency, todo_id: int = Path(ge=1)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def delete_todo(user: user_dependencies, db: db_dependency, todo_id: int = Path(ge=1)):
+    if not user:
+        raise HTTPException(status_code=404, detail="Authentication Failed")
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == int(user.get('id'))).first()
     if todo_model:
         db.delete(todo_model)
         db.commit()
